Remake database, fixed all relevants

+++ _database/procedure_course.sql
@@ -9,41 +9,59 @@ go
 create procedure ReadCourse @course_id int
 as
 begin
+	if not exists (
+		select 1
+		from Courses
+		where [course_id] = @course_id)
+	begin
+		select 'U_CID' as [check]
+		return;
+	end
+
 	select u.[user_id],
 		u.[user_full_name],
-		u.[address],
-		u.[birthdate],
-		u.[email],
-		u.phone_number,
+		u.[user_address],
+		u.[user_birthdate],
+		u.[user_email],
+		u.[user_phone_number],
 		c.[course_name], 
-		c.[course_description],
+		c.[course_short_description],
+		c.[course_full_description],
 		c.[course_price],
-		c.[duration],
-		m.[module_name],
-		m.[module_ordinal]
+		c.[course_duration],
+		c.[course_piority_index],
+		m.[module_ordinal],
+		m.[module_name]
 	from Courses c
 		left join Modules m on m.[course_id] = c.[course_id]
-		join Users u on u.[user_id] = c.[provider_id]
+		join Users u on u.[user_id] = c.[user_id]
 	where c.course_id = @course_id
+	order by m.[module_ordinal];
 end
 go
+-- exec ReadCourse 0
 ------------------------------------------------------------------------------------------------------------
 ------------------------------------------------------------------------------------------------------------
 if object_id('CreateCourse', 'P') is not null drop procedure CreateCourse;
 go
 create procedure CreateCourse 
 	@aid int, 
-	@course_name nvarchar(max), 
-	@course_description nvarchar(max),
+	@course_name nvarchar(max),
+	@course_short_description nvarchar(max),
+	@course_full_description nvarchar(max),
 	@course_price int,
-	@duration nvarchar(max)
+	@course_duration nvarchar(max)
 as
 begin
     declare @IsBanned BIT;
-    set @IsBanned = dbo.IsUserBanned(@aid, 'CreateCourse');
-    if @IsBanned = 1 return;
+	set @IsBanned = dbo.IsUserBanned(@aid, 'ViewUserCourses');
+    if @IsBanned = 1 
+	begin
+		select 'BANNED' as [check];
+	end
 
     declare @course_id int;
+	declare @user_id int;
 
     select top 1 @course_id = t1.course_id + 1
     from Courses t1
@@ -53,47 +71,136 @@ begin
 
     if @course_id is null select @course_id = isnull(max(course_id), 0) + 1 from Courses;
 
-    insert into Courses ([course_id], [course_name], [course_description], [course_price], [duration], [created_date], [provider_id])
-    values
-    (@course_id, @course_name, @course_description, @course_price, @duration, getdate(), @aid);
+	select @user_id = [user_id]
+	from Users
+	where [authentication_id] = @aid
 
-    select 'TRUE' as [check]
-    from Courses
-    where [course_id] = @course_id
-        and [course_name] = @course_name
-        and [course_description] = @course_description
-        and [duration] = @duration
-        and [provider_id] = @aid;
+	if (@user_id is null)
+	begin
+		select 'U_UID' as [check]
+		return;
+	end
+
+	if not exists (
+		select 1
+		from Authentications auth join Authorizations authz on authz.authorization_id = auth.authorization_id
+		where auth.authentication_id = @aid and authz.role = 'NAV_ESP')
+	begin
+		select 'U_ROLE' as [check]
+		return;
+	end
+
+	if exists (
+		select 1
+		from Courses
+		where [course_name] = @course_name)
+	begin
+		select 'EXISTED' as [check]
+		return;
+	end
+
+	declare @resource_url nvarchar(max) = '/courses/_' + cast(@course_id as nvarchar);
+
+    insert into Courses ([course_id],
+		[course_name],
+		[course_short_description],
+		[course_full_description],
+		[course_price],
+		[course_duration], 
+		[course_created_date],
+		[course_resource_url],
+		[course_piority_index],
+		[course_status],
+		[user_id]
+	)
+    values
+    (@course_id, @course_name, @course_short_description, @course_full_description, @course_price, @course_duration, getdate(), @resource_url, 1, 1, @user_id);
+
+	if @@ROWCOUNT = 1
+	begin
+		select 'SUCCESSED' as [check];
+		return;
+	end
+
+	select 'FAILED' as [check];
 end
 go
+-- exec CreateCourse 2, 'Test course name', 'Test short decrisption', 'Test full decrisption', 300000, '3 months', 5
+-- exec ReadCourse 2
 ------------------------------------------------------------------------------------------------------------
 ------------------------------------------------------------------------------------------------------------
 if object_id('UpdateCourse', 'P') is not null drop procedure UpdateCourse;
 go
-create procedure UpdateCourse @aid int, @course_id int, @new_course_name nvarchar(max), @new_course_description nvarchar(max), @new_duration nvarchar(max)
+create procedure UpdateCourse 
+	@aid int,
+	@course_id int,
+	@course_name nvarchar(max),
+	@course_short_description nvarchar(max),
+	@course_full_description nvarchar(max),
+	@course_price int,
+	@course_duration nvarchar(max),
+	@course_status bit
 as
 begin
 	declare @IsBanned BIT;
-	set @IsBanned = dbo.IsUserBanned(@aid, 'UpdateCourse');
-    if @IsBanned = 1 return;
+	set @IsBanned = dbo.IsUserBanned(@aid, 'ViewUserCourses');
+    if @IsBanned = 1 
+	begin
+		select 'BANNED' as [check];
+	end
+
+	declare @user_id int;
+
+	select @user_id = [user_id]
+	from Users
+	where [authentication_id] = @aid
+
+	if (@user_id is null)
+	begin
+		select 'U_UID' as [check]
+		return;
+	end
+
+	if not exists (
+		select 1
+		from Courses
+		where [course_id] = @course_id)
+	begin
+		select 'U_CID' as [check]
+		return;
+	end
+
+	if not exists (
+		select 1
+		from Authentications auth join Authorizations authz on authz.authorization_id = auth.authorization_id
+		where auth.authentication_id = @aid and authz.role = 'NAV_ESP')
+	begin
+		select 'U_ROLE' as [check]
+		return;
+	end
 
 	update Courses
-	set [course_name] = @new_course_name,
-		[course_description] = @new_course_description,
-		[duration] = @new_duration
-	where [course_id] = @course_id
-		and [provider_id] = @aid
+	set [course_name] = @course_name,
+		[course_short_description] = @course_short_description,
+		[course_full_description] = @course_full_description,
+		[course_price] = @course_price,
+		[course_duration] = @course_duration, 
+		[course_status] = @course_status
+	where [user_id] = @user_id
+		and [course_id] = @course_id
+	
 
-	select 'TRUE' as [check]
-	from Courses
-	where [course_id] = @course_id
-		and [provider_id] = @aid
-		and [course_name] = @new_course_name
-		and [course_description] = @new_course_description
-		and [duration] = @new_duration
+	if @@ROWCOUNT = 1
+	begin
+		select 'SUCCESSED' as [check];
+		return;
+	end
+
+	select 'FAILED' as [check];
 end
 go
-
+-- exec UpdateCourse 1, 2, 'Test course name1', 'Test short decrisption1', 'Test full decrisption1', 300000, '3 months', 1
+-- exec ReadCourse 2
 ------------------------------------------------------------------------------------------------------------
 ------------------------------------------------------------------------------------------------------------
 if object_id('DeleteCourse', 'P') is not null drop procedure DeleteCourse;
@@ -102,36 +209,89 @@ create procedure DeleteCourse @aid int, @course_id int
 as
 begin
 	declare @IsBanned BIT;
-	set @IsBanned = dbo.IsUserBanned(@aid, 'DeleteCourse');
-    if @IsBanned = 1 return;
+	set @IsBanned = dbo.IsUserBanned(@aid, 'ViewUserCourses');
+    if @IsBanned = 1 
+	begin
+		select 'BANNED' as [check];
+	end
+
+	declare @user_id int;
+
+	select @user_id = [user_id]
+	from Users
+	where [authentication_id] = @aid
+
+	if (@user_id is null)
+	begin
+		select 'U_UID' as [check]
+		return;
+	end
+
+	if not exists (
+		select 1
+		from Courses
+		where [course_id] = @course_id)
+	begin
+		select 'U_CID' as [check]
+		return;
+	end
+
+	if not exists (
+		select 1
+		from Authentications auth join Authorizations authz on authz.authorization_id = auth.authorization_id
+		where auth.authentication_id = @aid and authz.role = 'NAV_ESP')
+	begin
+		select 'U_ROLE' as [check]
+		return;
+	end
 
 	delete from Courses
 	where [course_id] = @course_id
-		and [provider_id] = @aid
+		and [user_id] = @user_id
 
-	select 'TRUE' as [check]
-	from Courses
-	where [course_id] = @course_id;
+	if @@ROWCOUNT = 1
+	begin
+		select 'SUCCESSED' as [check];
+		return;
+	end
+
+	select 'FAILED' as [check];
 end
 go
+-- exec DeleteCourse 1, 2
 ------------------------------------------------------------------------------------------------------------
 ------------------------------------------------------------------------------------------------------------
-if object_id('ViewUserCourses', 'P') is not null drop procedure ViewUserCourses;
+if object_id('ReadUserCourses', 'P') is not null drop procedure ReadUserCourses;
 go
-create procedure ViewUserCourses @aid int
+create procedure ReadUserCourses @aid int
 as
 begin
 	declare @IsBanned BIT;
 	set @IsBanned = dbo.IsUserBanned(@aid, 'ViewUserCourses');
-    if @IsBanned = 1 return;
+    if @IsBanned = 1 
+	begin
+		select 'BANNED' as [check];
+	end
 
-	select [course_id], [course_name], [course_price], [duration], [created_date]
+	declare @user_id int;
+
+	select @user_id = [user_id]
+	from Users
+	where [authentication_id] = @aid;
+
+	select [course_id],
+		[course_name],
+		[course_short_description],
+		[course_full_description],
+		[course_price],
+		[course_duration],
+		[course_created_date],
+		[course_status]
 	from Courses
-	where [provider_id] = @aid
+	where [user_id] = @user_id
 end
 go
-------------------------------------------------------------------------------------------------------------
-------------------------------------------------------------------------------------------------------------
+-- exec ViewUserCourses 1
 ------------------------------------------------------------------------------------------------------------
 ------------------------------------------------------------------------------------------------------------
 grant execute on dbo.ReadCourse to [NAV_GUEST];
